doc: Make all tutorial links track "latest" in master branch
While the master branch is in development state for the next stable branch,
its links should point to the "latest" docs branch, to ensure that users of
the unstable builds are linked to the relevant documentation.

Those links could be switched to stable branch subdomains before branching
off for a new major or minor release, to start tracking the frozen stable
docs branches.

See discussion in #29104.